# ideas and structures/openapi.frontend.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DAMN 前端所需的后端 API",
    "version": "0.2.0",
    "description": "自适应任务规划器前端（frontend/）当前依赖与将要依赖的全部 HTTP 接口。\n\n约定：\n- API 前缀 `/api/v1`，健康检查在根路径 `/health`，**不在** `/api/v1` 下。\n- 除 `/health`、`/auth/*` 外都需要 `Authorization: Bearer <JWT>`。\n- 所有错误统一返回 `{ code, message, detail }`。\n- 每个操作带 `x-implementation-status`：`existing` = 后端已有；`to-add` = 需要后端新增。\n- 日期用 `YYYY-MM-DD`；时间用 `HH:MM:SS`；时间戳用 ISO 8601。",
    "contact": { "name": "前端 | docs/design/backend-api-gaps.md" }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "开发环境（局域网）"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "tags": ["plans"],
        "summary": "计划列表",
        "description": "返回当前用户的所有计划版本；前端取 `status=active` 的那一版。",
        "x-implementation-status": "existing",
        "responses": {
          "200": {
            "description": "计划列表",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PlanListItem" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/plans/generate": {
      "post": {
        "tags": ["plans"],
        "summary": "生成计划（同步）",
        "description": "由目标直接生成并落库。前端的「拆解确认」在后端 draft 接口就绪前会回退到这里。同步返回完整 plan，不要做 SSE 假进度。",
        "x-implementation-status": "existing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PlanGenerateRequest" } }
          }
        },
        "responses": {
          "202": {
            "description": "生成完成（含 plan）",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlanGenerateResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
    "/api/v1/plans/decompose": {
      "post": {
        "tags": ["plans"],
        "summary": "批量拆解（目标页「拆解」）",
        "description": "把前端攒好的多条待拆解任务交给后端智能体，先返回一份**不落库**的草稿（按天分组），供只读预览。同一 draft_id 带新的 feedback 再次调用即「重新生成」。核心拆解由后端接智能体完成。",
        "x-implementation-status": "to-add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DecomposeRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "拆解草稿",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DecomposeResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
    "/api/v1/plans/confirm": {
      "post": {
        "tags": ["plans"],
        "summary": "确认拆解草稿（落库）",
        "description": "用户在预览里点「就先这样」时调用，把草稿落成正式计划并返回 PlanRead，前端直接应用。",
        "x-implementation-status": "to-add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmDraftRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "新计划",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PlanRead" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/plans/{plan_id}": {
      "get": {
        "tags": ["plans"],
        "summary": "计划详情",
        "description": "含该版本的全部任务与该计划涉及的目标。前端据此渲染待办时间轴。",
        "x-implementation-status": "existing",
        "parameters": [{ "$ref": "#/components/parameters/PlanId" }],
        "responses": {
          "200": {
            "description": "计划详情",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PlanRead" } }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/plans/{plan_id}/changes": {
      "get": {
        "tags": ["plans"],
        "summary": "计划变更记录（待办右上「计划变更」）",
        "description": "返回每次重排改了**未来哪几天**、每天新增/移动/移除几条、以及原因。触发者可能是反馈自动重排或手动重排。需后端在 `ReplanService.replan` 时把旧/新 `(scheduled_date, start_time, end_time)` 的 date diff 一起写入 ReplanEvent。",
        "x-implementation-status": "to-add",
        "parameters": [{ "$ref": "#/components/parameters/PlanId" }],
        "responses": {
          "200": {
            "description": "变更记录（按时间倒序）",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PlanChangeRead" } }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/v1/plans/{plan_id}/replan": {
      "post": {
        "tags": ["plans"],
        "summary": "手动重排",
        "description": "创建新版本（不可变，旧版标 superseded），复制未完成任务并重新排期。冷却中返回 409。",
        "x-implementation-status": "existing",
        "parameters": [{ "$ref": "#/components/parameters/PlanId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ReplanRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "重排结果",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReplanResponse" } }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/plans/{plan_id}/replan/eligibility": {
      "get": {
        "tags": ["plans"],
        "summary": "重排资格（冷却）",
        "description": "per-user 24h 冷却；前端用它决定是否允许再次重排并提示 `next_eligible_at`。",
        "x-implementation-status": "existing",
        "parameters": [{ "$ref": "#/components/parameters/PlanId" }],
        "responses": {
          "200": {
            "description": "资格",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReplanEligibilityRead" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plans/generation/{job_id}/events": {
      "get": {
        "tags": ["plans"],
        "summary": "生成进度（SSE）",
        "description": "事件名与图节点同名：`goal_analysis` / `theoretical_analysis` / `user_situation_analysis` / `plan_generation` / `rule_validation` / `plan_repair`，外加 `completed`。注意：当前前端不用 SSE（generate 同步返回）。",
        "x-implementation-status": "existing",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "事件流",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "PlanId": {
        "name": "plan_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未认证（401）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "Forbidden": {
        "description": "无权访问该资源（403）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "NotFound": {
        "description": "资源不存在（404）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "Conflict": {
        "description": "冲突（409）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "Validation": {
        "description": "参数校验失败（422）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "description": "统一错误体",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "机器可读错误码，如 `not_found` / `conflict` / `replan_not_eligible` / `validation_error`"
          },
          "message": { "type": "string", "description": "给人看的说明" },
          "detail": { "nullable": true, "description": "可选的补充信息（对象或数组）" }
        }
      },
...
      "GoalRead": {
        "type": "object",
        "required": ["id", "title", "goal_type", "priority", "status"],
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "goal_type": { "type": "string", "enum": ["long_term", "short_term", "project", "habit", "other"] },
          "priority": { "type": "integer", "minimum": 1, "maximum": 4 },
          "status": { "type": "string" }
        }
      },
...
      "GoalCreate": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "goal_type": { "type": "string", "enum": ["long_term", "short_term", "project", "habit", "other"], "default": "other" },
          "deadline": { "type": "string", "format": "date-time", "nullable": true },
          "priority": { "type": "integer", "minimum": 1, "maximum": 4, "default": 2 },
          "estimated_minutes": { "type": "integer", "nullable": true },
          "subject": { "type": "string", "nullable": true },
          "task_type": { "type": "string", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["active", "draft"],
            "default": "active",
            "description": "建待拆解项时传 draft"
          }
        }
      },
...
      "PlanRead": {
        "type": "object",
        "required": ["id", "user_id", "version", "status", "start_date", "end_date", "created_at", "tasks", "goals"],
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "version": { "type": "integer" },
          "status": { "type": "string", "enum": ["active", "superseded", "draft", "completed", "archived"] },
          "title": { "type": "string", "nullable": true },
          "start_date": { "type": "string", "format": "date" },
          "end_date": { "type": "string", "format": "date" },
          "parent_plan_id": { "type": "integer", "nullable": true },
          "confidence": { "type": "number", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "tasks": { "type": "array", "items": { "$ref": "#/components/schemas/TaskRead" } },
          "goals": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GoalRead" },
            "description": "由该计划任务的 goal_id 反查；前端只用它做分组标题"
          }
        }
      },
      "PlanListItem": {
        "type": "object",
        "required": ["id", "version", "status", "start_date", "end_date", "created_at", "task_count", "completed_count"],
        "properties": {
          "id": { "type": "integer" },
          "version": { "type": "integer" },
          "status": { "type": "string" },
          "title": { "type": "string", "nullable": true },
          "start_date": { "type": "string", "format": "date" },
          "end_date": { "type": "string", "format": "date" },
          "confidence": { "type": "number" },
          "created_at": { "type": "string", "format": "date-time" },
          "task_count": { "type": "integer" },
          "completed_count": { "type": "integer" }
        }
      },
      "PlanGenerateRequest": {
        "type": "object",
        "required": ["goals"],
        "properties": {
          "goals": { "type": "array", "minItems": 1, "items": { "$ref": "#/components/schemas/GoalCreate" } },
          "start_date": { "type": "string", "format": "date", "nullable": true },
          "end_date": { "type": "string", "format": "date", "nullable": true },
          "plan_title": { "type": "string", "nullable": true },
          "available_minutes_per_day": { "type": "integer", "minimum": 1, "maximum": 1440, "default": 480 },
          "daily_limit_minutes": { "type": "integer", "minimum": 1, "maximum": 1440, "default": 300 },
          "buffer_minutes": { "type": "integer", "minimum": 0, "maximum": 120, "default": 15 },
          "high_cognitive_max_per_day": { "type": "integer", "minimum": 1, "maximum": 10, "default": 2 },
          "user_profile": { "type": "object", "additionalProperties": true, "default": {} },
          "execution_weight": { "type": "number", "nullable": true }
        }
      },
      "PlanGenerateResponse": {
        "type": "object",
        "required": ["job_id", "status", "events_url"],
        "properties": {
          "job_id": { "type": "string" },
          "status": { "type": "string", "example": "completed" },
          "plan_id": { "type": "integer", "nullable": true },
          "events_url": { "type": "string" },
          "plan": { "allOf": [{ "$ref": "#/components/schemas/PlanRead" }], "nullable": true }
        }
      },
      "DecomposeRequest": {
        "type": "object",
        "required": ["goals"],
        "properties": {
          "goals": {
            "type": "array",
            "minItems": 1,
            "description": "待拆解清单（用户在目标页攒的多条）",
            "items": { "$ref": "#/components/schemas/DecomposeGoal" }
          },
          "draft_id": {
            "type": "string",
            "nullable": true,
            "description": "「重新生成」时带上同一 draft_id，后端据此保留上下文"
          },
          "feedback": {
            "type": "string",
            "nullable": true,
            "description": "用户对上一版预览的意见，如「晚点做」「少一点」「挪到周末」"
          }
        }
      },
      "DecomposeGoal": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "priority": { "type": "integer", "minimum": 1, "maximum": 3, "default": 2 },
          "deadline": { "type": "string", "format": "date", "nullable": true },
          "notes": { "type": "string", "description": "补充，拆解线索" }
        }
      },
...
      "DecomposeDay": {
        "type": "object",
        "required": ["date", "tasks"],
        "properties": {
          "date": { "type": "string", "format": "date" },
          "tasks": { "type": "array", "items": { "$ref": "#/components/schemas/DecomposeDraftTask" } }
        }
      },
      "DecomposeResponse": {
        "type": "object",
        "required": ["draft_id", "days"],
        "properties": {
          "draft_id": { "type": "string" },
          "days": { "type": "array", "items": { "$ref": "#/components/schemas/DecomposeDay" } }
        }
      },
      "ConfirmDraftRequest": {
        "type": "object",
        "required": ["draft_id"],
        "properties": { "draft_id": { "type": "string" } }
      },
...
      "ReplanEligibilityRead": {
        "type": "object",
        "required": ["eligible", "reason", "cooldown_hours"],
        "properties": {
          "eligible": { "type": "boolean" },
          "reason": { "type": "string" },
          "next_eligible_at": { "type": "string", "format": "date-time", "nullable": true },
          "last_replan_at": { "type": "string", "format": "date-time", "nullable": true },
          "cooldown_hours": { "type": "integer", "example": 24 }
        }
      },
...
      "ReplanTriggerType": {
        "type": "string",
        "enum": ["manual", "feedback_triggered", "scheduled", "system"]
      },
      "ReplanRequest": {
        "type": "object",
        "properties": {
          "trigger_type": { "$ref": "#/components/schemas/ReplanTriggerType" },
          "reason": { "type": "string", "nullable": true },
          "from_date": { "type": "string", "format": "date", "nullable": true }
        }
      },
      "ReplanResponse": {
        "type": "object",
        "required": ["plan_id", "old_version", "new_version", "changed_task_ids", "reason", "trigger_type"],
        "properties": {
          "plan_id": { "type": "integer" },
          "old_version": { "type": "integer" },
          "new_version": { "type": "integer" },
          "changed_task_ids": { "type": "array", "items": { "type": "integer" } },
          "reason": { "type": "string" },
          "trigger_type": { "$ref": "#/components/schemas/ReplanTriggerType" }
        }
      },
...
      "PlanChangeRead": {
        "type": "object",
        "required": ["id", "trigger_type", "reason", "old_version", "new_version", "created_at", "days"],
        "properties": {
          "id": { "type": "integer" },
          "trigger_type": { "$ref": "#/components/schemas/ReplanTriggerType" },
          "reason": { "type": "string" },
          "old_version": { "type": "integer" },
          "new_version": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "days": { "type": "array", "items": { "$ref": "#/components/schemas/PlanChangeDayRead" } }
        }
      },
...
    }
  }
}